quick test with a django-multiselectfield.   works but with some caveats.
I changed the stored values to strings so they go in and out without any controvesy.
I also note that if you want to query by these values, multiselect has a couple of wrinkles:
   *  searches are case insensitive
   *  searches will return partial matches
   *  additions through the code are not validated.

Take for example:
   Snippet.objects.filter(mselect__contains='3')  # will return snippets with '3.5'
   Snippet.objects.filter(mselect__contains='ca')  # will return snippets with 'CA'
   s = Snippet.objects.first()
   s.mselect.append('steve')   # it's just a list
   s.save()  # no exception thrown here, either
   Snippet.objects.filter(mselect__contains='Eve')  # matches 'steve',  there be dragons

The good news is that DRF does validate, so trying to PUT or PATCH data not in the choices
does return a 400  Bad Request.